Option for using this addon when kodi isn't playing
Add the option for using the subscenter addon when kodi is no playing
any media, by using the Manual Search button in the Subtitles Dialog.
* Open the Subtitles Dialog outside the player can be done by add it in
the Keymap xml files.

* This patch by #burekas doesn't change the noraml operation as it today
when kodi is playing media.
* This fix have been tested by me (burekas)
* Follow the #burekas comments where the changes have been done.
The mainly changes are by add a condition where kodi playing or not
playing: if xbmc.Player().isPlaying()