Addition to PR #7 - Title will be taken from the focused item (#15)
* Fix Download exception when kodi isn't playing

* Fix the error exception from kodi when downloading subtitle is done
when kodi isn't playing (Added an if conditition in mirror_sub function)
Note: When the user will press on a subtitle, the subtitle will still be
downloaded to the temp folder in the addon folder in userdata (As it
done currently for normal downloading when kodi is playing), after that
the search subtitles dialog will be closed. But this is not should be
relevant to the user for the main purpose of using it only for watching
the subtitles result list.

In Addition:
* Set a new local variable __isKodiPlaying__ instead of writing
xbmc.Player().isPlaying() on each place.
* Fix speeling "Serach for" to "Search for..."

* Fix Download exception when kodi isn't playing

* Fix the error exception from kodi when downloading subtitle is done
when kodi isn't playing (Added an if conditition in mirror_sub function)
* Note: When the user will press on a subtitle, the subtitle will still
be downloaded to the temp folder in the addon folder in userdata (As it
done currently for normal downloading when kodi is playing), after that
the search subtitles dialog will be closed. But this is not should be
relevant to the user for the main purpose of using it only for watching
the subtitles result list.

In Addition:
* Set a new local variable __isKodiPlaying__ instead of writing
xbmc.Player().isPlaying() on each place.
* Fix speeling "Serach for" to "Search for..."

* Revert "Fix Download exception when kodi isn't playing"

This reverts commit 060d0e8604bbd94aef25b11d59e10048c044d327.

* Fix condition

Fix condition

* Critical Fix - If Condition when title is empty.

Salts and Local ontent for example gives empty title from the player,
while exodus and specto gives a title. So in Exodus case it always goes
to the "else" and didn't show results.
The if for kodi is playing must be inside the if for empty title. So
when the title is empty it takes the right title. The "else" must be
inside because 2 addons behave different.

I check this and now it's ok.

* Critical Fix - If condition when title is empty

Salts and Local content for example gives empty title from the player,
while exodus and specto gives a title. So currently in Exodus case it
always goes to the "else" with the "fake" title and didn't show results.
When with salts it go into the right place and fix the title.

The fix:
Basically, put the "fake" title few lines before, in the first condition
if kodi is playing (Instead of the empty title as it now), and bring
back the if condition when title is empty as it was before (without
checking there if kodi is playing). But I didn't check this like that.

I check this and now it's ok.

* Critical Fix - Regarding If condition when title is empty

Salts and Local content for example gives empty title from the player,
while exodus and specto gives a title. So currently in Exodus case it
always goes to the "else" with the "fake" title and didn't show results.
When with salts it go into the right place and fix the title.

The fix:
Basically, put the "fake" title few lines before, in the first condition
if kodi is playing (Instead of the empty title as it now), and bring
back the if condition when title is empty as it was before (without
checking there if kodi is playing). But I didn't check this like that.

I check this and now it's ok.

* Critical Fix - If condition when title is empty



Salts, Quasar and Local content for example gives empty title from the player, while exodus and specto gives a title. So currently in Exodus case it always goes to the "else" with the "fake" title and didn't show results.
When with salts it go into the right place and fix the title.

The fix:
Basically, put the "fake" title few lines before, in the first condition if kodi is playing (Instead of the empty title as it now), and bring back the if condition when title is empty as it was before (without checking there if kodi is playing).

I check this and now it's ok.

* New IMDB_ID method

* Version 4.1.2

* Improvment for PR #7 - Using when kodi isn't playing

Improvement for the option of using the addon when kodi is not playing:
* Now the title will be taken from the focused item when it has details
(For example: Kodi Library, Exodus. But with Salts, Quasar, Specto this
will still not work until their developers will fix it)
* If it doesn't have details, it will remain "SearchFor" in order to
present "No Subtitles Found" result

* Fix comment

* Revert "Fix comment"

This reverts commit a4c7aac5e3cc92596a47a9b06b7baa465ed40f88.

* Revert "Improvment for PR #7 - Using when kodi isn't playing"

This reverts commit 03a867bc55ab5419914cc127862d571744bb2a5b.

* Improvement for PR #7 - Using the addon when Kodi isn't playing

Improvement for the option of using the addon when kodi is not playing:
* Now the title will be taken from the focused item when it has details
(For example: Kodi Library, Exodus. But with Salts, Quasar, Specto this
will still not work until their developers will fix it)
* If it doesn't have details, it will remain "SearchFor" in order to
present "No Subtitles Found" result, so you can typing it manually in
the "Manual Search"

* Addition to PR #7 - Title will be taken from the focused item

Improvement for the option of using the addon when kodi is not playing:
* Now the title will be taken from the focused item when it has details.
* If it doesn't have details, it will remain "SearchFor" in order to
present "No Subtitles Found" result, so you can typing it manually in
the "Manual Search"

* Update changelog.txt